LayoutTests/platform/mac-mojave/fast/text/international/system-language/navigator-language/navigator-language-es-expected.txt

FAIL navigator.language should be es. Was es-es.
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
